ImprovedBranch/requirements.txt
certifi==2024.12.14
charset-normalizer==3.4.0
colorama==0.4.6
contourpy==1.2.1
cycler==0.12.1
easydict==1.13
einops==0.4.1
filelock==3.16.1
fonttools==4.55.3
fsspec==2024.2.0
fvcore==0.1.5.post20221221
idna==3.10
imageio==2.36.1
iopath==0.1.10
Jinja2==3.1.5
kiwisolver==1.4.7
MarkupSafe==3.0.2
matplotlib==3.8.4
mpmath==1.3.0
networkx==3.4.2
numpy==1.26.4
packaging==24.2
pandas==2.2.3
pillow==11.0.0
portalocker==3.0.0
pyparsing==3.2.0
python-dateutil==2.9.0.post0
pytz==2024.2
PyWavelets==1.6.0
PyYAML==6.0.2
requests==2.32.3
scikit-image==0.19.3
scipy==1.13.1
six==1.17.0
sympy==1.13.1
tabulate==0.9.0
termcolor==2.5.0
tifffile==2024.12.12
timm==0.5.4
torchaudio
tqdm==4.67.1
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.3.0
yacs==0.1.8
torchxrayvision
--index-url https://download.pytorch.org/whl/cu121
--extra-index-url https://pypi.org/simple
scikit-learn
torch==2.2.2
torchvision==0.17.2
seaborn